chore: hide request history and refactor request history
